// src/redux/api/reviews/types.d.ts
namespace REVIEWS {
  // Клиент
  export interface Client {
    id: number;
    first_name: string;
    last_name: string;
    user_picture: string | null;
    from_user: string;
  }

  // Изображение
  export interface ReviewImage {
    id: number;
    image: string;
  }
// ...
  export interface ReplyHotelReview {
    id: number
    user: Client
    comment: string
    created_date: string
  }

  export interface Review {
    id: number;
    entityId: number | string; 
    client: Client;
    comment: string;
    rating: number
    reviewImages: ReviewImage[];
    createdAt: string; 
    replyReviews: ReplyHotelReview[]
  }

  export interface StaticReview {
    id: number;
    name?: string; 
    avgRating: number; 
    ratingCount: number;
    excellent: number;
    good: number;
    notBad: number;
    bad: number;
    terribly: number;
  }
// ...
}